refactor overlay synchronization, synchronizedOverlay now extends ol.Overlay
